+++ include/server/protocol.hpp
@@ -2,7 +2,7 @@
 #include "sqlite3.h"
 
 #define BUFFER_SIZE 512
-#define SendFn std::function<void(char *, int)>
+#define SendFn std::function<void(char *, size_t)>
 
 namespace ti {
 namespace server {
@@ -12,9 +12,9 @@ class Client {
   public:
     virtual ~Client() = default;
     void initialize(SendFn fn);
-    void send(char *content, int len) const;
+    void send(char *content, size_t len) const;
     virtual void on_connect(sockaddr_in addr) = 0;
-    virtual void on_message(char *content, int len) = 0;
+    virtual void on_message(char *content, size_t len) = 0;
     virtual void on_disconnect() = 0;
 };
 class Server {
@@ -27,7 +27,7 @@ class Server {
   public:
     Server(std::string addr, short port);
     ~Server();
-    void send(SocketFd clientfd, char *data, int len);
+    void send(SocketFd clientfd, char *data, size_t len);
     virtual Client *on_connect(sockaddr_in addr) = 0;
     void start();
     void stop();
